+++ changelog.md
@@ -9,6 +9,8 @@
 * Refactored enchanting feature tooltip
   * It was becoming spaghetti and there were untranslatable strings
   * Also fixed Base Enchantability showing on non-enchantable items
+* Fixed Curses not being appliable to enchanted items
+  * You still have to disenchant the item to use the bonus enchantability, but you do you
 
 ## 1.8.7
 * Fixed Gravity Defying applying fall damage reduction even without the enchantment